product, purchase 자료형 수정 - productPrice(double -> int), purchaseDate(String -> Date), purchase(String -> int)

+++ notalone/src/main/java/com/webprojectv1/notalone/product/Product.java
@@ -17,7 +17,7 @@ public class Product {
     private String productCategory;
 
     @Column(nullable = false)
-    private double productPrice;
+    private int productPrice;
 
     @Column(nullable = false)
     private double productRating;
